requirements.txt
PySide6>=6.6.0
QDarkStyle
networkx>=3.2
pydot>=2.0.0
requests
httpx
pyyaml
loguru
rich
androguard
capstone
keystone-engine
unicorn
frida-tools
mitmproxy
pycryptodome
pefile
lief